pyproject.toml
[tool.poetry]
name = "referral-system-fastapi"
version = "0.1.0"
description = ""
authors = ["Kettariec <kettariec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.5"}
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic = "^2.10.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
